+++ packages/graphql-transformer-common/src/ResourceConstants.ts
@@ -35,6 +35,7 @@ export class ResourceConstants {
         DynamoDBBillingMode:   'DynamoDBBillingMode',
         DynamoDBModelTableReadIOPS: 'DynamoDBModelTableReadIOPS',
         DynamoDBModelTableWriteIOPS: 'DynamoDBModelTableWriteIOPS',
+        DynamoDBEnablePointInTimeRecovery: 'DynamoDBEnablePointInTimeRecovery',
 
         // Elasticsearch
         ElasticsearchDomainName: 'ElasticSearchDomainName',
@@ -50,10 +51,6 @@ export class ResourceConstants {
 
         // Auth
         AuthCognitoUserPoolId: 'AuthCognitoUserPoolId',
-        AuthCognitoUserPoolName: 'AuthCognitoUserPoolName',
-        AuthCognitoUserPoolMobileClientName: 'AuthCognitoUserPoolMobileClientName',
-        AuthCognitoUserPoolJSClientName: 'AuthCognitoUserPoolJSClientName',
-        AuthCognitoUserPoolRefreshTokenValidity: 'AuthCognitoUserPoolRefreshTokenValidity'
     }
     public static MAPPINGS = {}
     public static CONDITIONS = {
@@ -61,11 +58,11 @@ export class ResourceConstants {
         HasEnvironmentParameter: 'HasEnvironmentParameter',
 
         // DynamoDB
-        ShouldUsePayPerRequestBilling:   'ShouldUsePayPerRequestBilling',
+        ShouldUsePayPerRequestBilling: 'ShouldUsePayPerRequestBilling',
+        ShouldUsePointInTimeRecovery: 'ShouldUsePointInTimeRecovery',
 
         // Auth
-        AuthShouldCreateUserPool: 'AuthShouldCreateUserPool',
-        APIKeyExpirationEpochIsNotNegOne: 'APIKeyExpirationEpochIsNotNegOne',
+        ShouldCreateAPIKey: 'ShouldCreateAPIKey',
         APIKeyExpirationEpochIsPositive: 'APIKeyExpirationEpochIsPositive',
     }
     public static OUTPUTS = {
@@ -78,6 +75,8 @@ export class ResourceConstants {
         // Elasticsearch
         ElasticsearchStreamingLambdaIAMRoleArn: 'ElasticsearchStreamingLambdaIAMRoleArn',
         ElasticsearchAccessIAMRoleArn: 'ElasticsearchAccessIAMRoleArn',
+        ElasticsearchDomainArn: 'ElasticsearchDomainArn',
+        ElasticsearchDomainEndpoint: 'ElasticsearchDomainEndpoint',
 
         // Auth
         AuthCognitoUserPoolIdOutput: 'AuthCognitoUserPoolIdOutput',