+++ packages/amplify-category-auth/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-auth@1.0.6...amplify-category-auth@1.0.7) (2019-03-22)
+
+
+### Bug Fixes
+
+* **amplify-category-auth:** use right response signal of cfn-response ([572ca45](https://github.com/aws-amplify/amplify-cli/commit/572ca45))
+
+
+
+
+
 ## [1.0.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-auth@1.0.5...amplify-category-auth@1.0.6) (2019-02-26)